pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "guardrail-compliance-engine"
version = "0.1.0"
description = "Terraform-first compliance scanning with Bedrock Guardrails Automated Reasoning support"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [{ name = "Varad More" }]
dependencies = [
  "boto3>=1.35.0",
  "PyYAML>=6.0.1",
  "python-hcl2>=4.3.5",
  "rich>=13.7.0",
  "typer>=0.12.3",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.2.0",
]

[project.scripts]
guardrail = "guardrail_compliance.cli:app"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]
